# `Std.parseInt()` fixes and optimisations