--- a/std/js/_std/Std.hx
+++ b/std/js/_std/Std.hx
@@ -54,17 +54,24 @@ import js.Syntax;
 
 	@:pure
 	public static function parseInt(x:String):Null<Int> {
-		if(x != null) {
-			for(i in 0...x.length) {
-				var c = StringTools.fastCodeAt(x, i);
-				if(c <= 8 || (c >= 14 && c != ' '.code && c != '-'.code)) {
-					var nc = StringTools.fastCodeAt(x, i + 1);
-					var v = js.Lib.parseInt(x, (nc == "x".code || nc == "X".code) ? 16 : 10);
-					return Math.isNaN(v) ? null : cast v;
-				}
+		#if (js_es >= 5)
+		final v = js.Lib.parseInt(x);
+		#else
+		// before ES5, octal was supported in some implementations, so we need to explicitly use base 10 or 16
+		if (x == null)
+			return null;
+		var v:Float = Math.NaN;
+		for (i => c in StringTools.keyValueIterator(x)) {
+			if ((c <= 8 || c >= 14) && !(c == ' '.code || c == '-'.code || c == '+'.code)) {
+				final nc = js.Syntax.code("{0}[{1}]", x, i + 1);
+				v = js.Lib.parseInt(x, c == '0'.code && (nc == "x" || nc == "X") ? 16 : 10);
+				break;
 			}
 		}
-		return null;
+		#end
+		if (Math.isNaN(v))
+			return null;
+		return cast v;
 	}
 
 	public static inline function parseFloat(x:String):Float {