--- a/std/php/_polyfills.php
+++ b/std/php/_polyfills.php
@@ -60,4 +60,12 @@ namespace { //Namespace declaration is required because this file is included un
 		}
 	}
 
+	/**
+	 * @see https://www.php.net/manual/en/function.str-starts-with.php
+	 */
+	if (!function_exists('str_starts_with')) {
+		function str_starts_with($str, $start) {
+    		return (@substr_compare($str, $start, 0, strlen($start))==0);
+		}
+	}
 }