--- a/tests/unit/src/unit/issues/Issue9198.hx
+++ b/tests/unit/src/unit/issues/Issue9198.hx
@@ -1,0 +1,9 @@
+package unit.issues;
+
+import unit.Test;
+
+class Issue9198 extends Test {
+	public function test() {
+		eq(null, Std.parseInt("axolotl"));
+	}
+}