Improvements to parseInt implementation on various targets, along with unit tests.

Unit tests for parseInt and parseFloat have been organised and labelled. Documentation has also been updated to be more specifc about edge cases.

### Dependent PRs
- HaxeFoundation/neko#258
- HaxeFoundation/hashlink#532
- HaxeFoundation/hxcpp#988

### Fixes
- Accept `+` in front of hexadecimal number (Javascript, Php, Hashlink, Neko)
- Accept `-` in front of hexadecimal number (Neko) #10545
- Do not allow multiple signs, or space between sign and number (Eval)
- Allow trailing text after Hexadecimal numbers (Java, JVM, C#, Python, C++, Hashlink, Neko)
- Ignore all leading whitespace in parseInt (C++, Hashlink, Neko, Lua, Php)
- Ignore all leading whitespace in parseFloat (Eval)
- Fix hexadecimal check (Javascript) Fixes #9198
- Do not crash when hexadecimal number has two signs or hexadecimal prefix is empty (Lua)

### Optimisations
Looking through some older issues, I saw that performance was a concern for some, as making guarantees about specific edge cases could require extra overhead. In light of this, I have made sure to avoid affecting performance. The parseInt implementation on most targets has been left as fast if not faster than it were before. I did some benchmarks which can be seen below:
<details>
<summary>Results</summary>

Random integer strings were generated, with random whitespace etc. Different targets had to have different sample sizes due to memory limitations.

The most significant optimisations were on C#, Java, JVM, and Python. Other targets have slightly better or roughly the same performance as before. Hashlink and C++ are exceptions, however, this was unavoidable as Hashlink was lacking proper whitespace checks, and C++ lacked support for + in front of hexadecimals.
```
- Python

Samples: 2500000
Old: 65.13595397600147 seconds
New: 14.160146241999428 seconds

- C#

Samples: 10000000
Old: 20.2922110557556 seconds
New: 3.04013705253601 seconds

- Java

Samples: 10000000
Old: 4.921999931335449 seconds
New: 1.3929998874664307 seconds

- JVM

Samples: 10000000
Old: 4.951000213623047 seconds
New: 1.125 seconds

- Hashlink

Samples: 10000000
Old: 1.38132810592651 seconds
New: 1.51130199432373 seconds

- PHP

Samples: 697500
Old: 0.42882704734802 seconds
New: 0.31195902824402 seconds

- Interp

Samples: 10000000
Old: 2.52953886985778809 seconds
New: 2.2584080696105957 seconds

- Lua

Samples: 697500
Old: 1.3906750679016 seconds
New: 1.3417611122131 seconds

- Neko

Samples: 10000000
Old: 3.3531289100647 seconds
New: 3.22968316078186 seconds

- Javascript

Samples: 2500000
Old: 5.183575558999554 seconds
New: 4.908442030999808 seconds

- C++

Samples: 6900000
Old: 12.0737991333008 seconds
New: 13.7655079364777 seconds

- C++IA

Samples: 6000000
Old: 1.7640278339386 seconds
New: 1.76914310455322 seconds

- Flash

Samples: 10000
Old: 0.008000000000000007 seconds
New: 0.008000000000000007 seconds

```
</details>