@Simn 2022-04-03
Nice one! Thanks a lot for your work once again.

The benchmarks are a bit shocking. I wonder where this difference actually comes from on e.g. JVM. But that depends on what exactly you're benchmarking.

Also I like this `hasIndex` pattern, very clean.