@tobil4sk 2022-04-03
Thanks!

For C#, JVM, Java, and Python, I think the main performance benefit is that before, there was one large while loop with a switch statement which performed every character check, and kept track of the state with variables. This meant that even if it had already started parsing digits, it was still doing checks for whitespaces and signs etc. before it checked if it was a digit, and it used variables to keep track of how it should handle these characters if digits have already been found. For example whitespace, is only ignored before any digits, but the check for it is always done at any point.

Now each part of the number is handled separately, so first the whitespace, then the sign, then the digits, up until a non digit character. This avoids the dependence on "state" variables to keep track of which part of the parsing process we are at, because this is done by the control flow of the code. This also means that fewer checks have to be done at once, because we are only checking on thing at a time.

Additionally, Python now uses array access to get characters instead of character codes, because from profiling it, it seemed the "ord()" function was taking up a fair amount of execution time.

The benchmarks were done on randomly generated numbers, which all had a random number of preceeding spaces, a randomly chosen sign (or not), then a series of either hexadecimal or decimal digits, and finally with a random bit of trailing text. I'm wondering now whether it would be good to see how it performs with just plain numbers (and maybe other more common cases). Theoretically it should still be faster but might be worth checking. I'll also zip up the benchmark code for reference, I just have to do some cleaning up first.