@tobil4sk 2022-04-08
After running the C++ benchmarks again the new version is a bit slower, however, this is because + is now handled properly in front of hexadecimal digits (more valid numbers parsed leads to a slower result because of the way I benchmarked it). I've changed the details in the original post.

Just for reference here is the code I used for benchmarking: [zipped.zip](https://github.com/HaxeFoundation/haxe/files/8455612/zipped.zip)